+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Upgraded to Unity 6000.0.56f1 and Burst 1.8.24.
+- `HexGameUI` and `HexMapEditor` keep track of selected cell via `HexCell` struct.
+
 ## [4.0.0] - 2025-03-30
 
 ### Changed